+++ src/com/tests/DivisionTest.java
@@ -22,9 +22,9 @@ public class DivisionTest
 	public void test_setId_getId()
 	{
 		tmpDiv.setId(10);
-		assertEquals(10, tmpDiv.getId());
+		assertEquals(10, tmpDiv.getId().intValue());
 		tmpDiv.setId(8);
-		assertEquals(8, tmpDiv.getId());
+		assertEquals(8, tmpDiv.getId().intValue());
 	}
 
 	@Test
@@ -40,18 +40,18 @@ public class DivisionTest
 	public void test_setHeadDivId_getHeadDivId()
 	{
 		tmpDiv.setHeadDivId(10);
-		assertEquals(10, tmpDiv.getHeadDivId());
+		assertEquals(10, tmpDiv.getHeadDivId().intValue());
 		tmpDiv.setHeadDivId(9);
-		assertEquals(9, tmpDiv.getHeadDivId());
+		assertEquals(9, tmpDiv.getHeadDivId().intValue());
 	}
 
 	@Test
 	public void test_setChiefId_getChiefId()
 	{
 		tmpDiv.setChiefId(10);
-		assertEquals(10, tmpDiv.getChiefId());
+		assertEquals(10, tmpDiv.getChiefId().intValue());
 		tmpDiv.setChiefId(9);
-		assertEquals(9, tmpDiv.getChiefId());
+		assertEquals(9, tmpDiv.getChiefId().intValue());
 	}
 
 	@Test
